requirements.txt
PyQt6
Pillow
imagehash
send2trash
